Seven Game
Game in which the player needs to enter a number. Be careful though, if you enter a 7, a number containing 7 or a multitude of 7, you have to drink.

This game is not fully finished. It needs to have the functionality that checks whether numbers are submitted in sequence (round 1 requires a submission of a 1.